magically fixed now?

+++ src/socket.cpp
@@ -232,7 +232,7 @@ wxString Socket::GetHandle()
 
     DWORD dwStatus = GetAdaptersInfo ( AdapterInfo, &dwBufLen); // Get info
 		if (dwStatus != NO_ERROR) return _T(""); // Check status
-    for (unsigned int i=0; i<std::min(6, AdapterInfo[0].AddressLength); i++)
+    for (unsigned int i=0; i<std::min( (unsigned int)6, (unsigned int)AdapterInfo[0].AddressLength); i++)
     {
         handle += TowxString(((unsigned int)AdapterInfo[0].Address[i])&255);
         if (i != 5) handle += _T(':');
